move numpy json encoder to common area where it can be referenced by 2 calls to json.dump
